+++ src/pages/home.jsx
@@ -1,10 +1,9 @@
 import React from "react";
-import LeftBar from "./test/left-bar";
 
 const Home = () => {
 
     return < >
-        <LeftBar />
+        Home
     </>
 }
 export default Home;
