--- src/pages/test/succesfly.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Succesfly = () => {
-  return (
-    <div>
-        
-    </div>
-  )
-}
-
-export default Succesfly
